task3/procrastinizer/requirements.txt
djangorestframework
django
psycopg2

